% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundances.R
\name{createAbuAgent}
\alias{createAbuAgent}
\title{Create Abundance Agent}
\usage{
createAbuAgent(talysStructurePath)
}
\arguments{
\item{talysStructurePath}{path to the structure/abundance directory of TALYS}
}
\value{
List with functions
        \tabular{ll}{
          \code{setAbundances} \tab TODO \cr
          \code{getAbundances(sym)} \tab
          get a data table with abundances of isotopes of an element \code{sym}, e.g. 'Fe'
        }
}
\description{
Constructs a closure with function to set and query abundances of elements.
}
