% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subent_handler_generic.R
\name{createDefaultSubentHandlerList}
\alias{createDefaultSubentHandlerList}
\title{Create List of Default Subentry Handler}
\usage{
createDefaultSubentHandlerList()
}
\value{
List of subentry handlers
}
\description{
Create List of Default Subentry Handler
}
