% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subent_handler_generic.R
\name{createSubentHandler}
\alias{createSubentHandler}
\title{Create Subentry Handler}
\usage{
createSubentHandler(subentHandlerList)
}
\value{
List with functions:
        \describe{
          \item{getName()}{returns signature of the general handler}
          \item{getInfo()}{returns additional information about handler}
          \item{canMap(subent)}{returns a bool indicating whether \code{subent} can be mapped}
          \item{needs(subent)}{returns data table with extended output specification, see ...}
          \item{map(subent)}{TODO}
        }
}
\description{
Create Subentry Handler
}
