% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subent_handler_ntot.R
\name{createSubentHandler_ntot}
\alias{createSubentHandler_ntot}
\title{Create Specific Subent Handler}
\usage{
createSubentHandler_ntot()
}
\value{
Returns closure with functions as given in [.].
        In addition, the closure contains the functions \code{setAbuAgent}.
}
\description{
Handles an EXFOR entry with (N,TOT) for a specific target
}
