% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subent_handler_ntot_nat.R
\name{createSubentHandler_ntot_nat}
\alias{createSubentHandler_ntot_nat}
\title{Create Specific Subentry Handler}
\usage{
createSubentHandler_ntot_nat()
}
\value{
Returns closure with functions as given in [.].
        In addition, the closure contains the functions \code{setAbuAgent}.
}
\description{
Handles an EXFOR entry with (N,TOT) and target in natural composition
}
